[FAB-4134] Consolidate unit-test support config
Peer code contains 5 different config.go as support configuration
for unit-tests in various packages. This changeset refactor and
consolidate them into 1 module.
