+++ events/producer/producer_test.go
@@ -39,7 +39,7 @@ import (
 	mmsp "github.com/hyperledger/fabric/common/mocks/msp"
 	"github.com/hyperledger/fabric/common/util"
 	"github.com/hyperledger/fabric/core/config"
-	"github.com/hyperledger/fabric/events"
+	coreutil "github.com/hyperledger/fabric/core/testutil"
 	"github.com/hyperledger/fabric/events/consumer"
 	"github.com/hyperledger/fabric/msp"
 	"github.com/hyperledger/fabric/msp/mgmt"
@@ -495,7 +495,7 @@ func TestMain(m *testing.M) {
 		os.Exit(-1)
 		return
 	}
-	events.SetupTestConfig()
+	coreutil.SetupTestConfig()
 	var opts []grpc.ServerOption
 	if viper.GetBool("peer.tls.enabled") {
 		creds, err := credentials.NewServerTLSFromFile(config.GetPath("peer.tls.cert.file"), config.GetPath("peer.tls.key.file"))
